requirements.txt
### dependency to be installed
tqdm
mlflow
pandas
numpy
PyYAML
Pillow
tensorflow